+++ cpp/oneapi/dal/algo/shortest_paths/backend/cpu/traverse_default_kernel_cpu.cpp
@@ -18,6 +18,12 @@
 
 namespace oneapi::dal::preview::shortest_paths::backend {
 
-template bool nrh_dispatcher<__CPU_TAG__>();
+template struct delta_stepping<__CPU_TAG__, std::int32_t>;
+
+template struct delta_stepping<__CPU_TAG__, double>;
+
+template struct delta_stepping_with_pred<__CPU_TAG__, std::int32_t>;
+
+template struct delta_stepping_with_pred<__CPU_TAG__, double>;
 
 } // namespace oneapi::dal::preview::shortest_paths::backend
